test(mga): add positive-coverage E2E for DesktopBadge via Tauri shim
Rename desktop-badge-absent-in-web.spec.ts → desktop-badge.spec.ts and
add a third test that injects a fake `window.__TAURI_INTERNALS__` via
`page.addInitScript` before the SPA bundle evaluates. Asserts the badge
renders with the mocked version/build/PR values.

This covers the gated-component DOM render path end-to-end without
needing `tauri-driver` / WebDriver wiring. The real Rust IPC bridge is
covered by the cross-platform CI build matrix.
